vehicle refact
